Switch call arguments to be a `call_arg list`
Summary:
I dislike `RestT` and intend to kill it. A spread argument is not just
a normal argument with a different type.

In this diff, I keep track of which arguments are normal arguments and which
ones are spreads. This forces us to deal with spread arguments in places where
previously we forgot they might exist.

Reviewed By: samwgoldman

Differential Revision: D4404830

fbshipit-source-id: 9e5f81fa730682e0724f48935c1ef128444324e1