Propagate any to ReactCreateElementT
Summary:
In D3285681 I made a CustomFunT for React.createElement. That custom fun waits
for the React component class to resolve, then does the appropriate check for
class-based components, functional components, and intrinsics.

However, if the component class resolves to `any`, the constraint would be
eagerly pruned in `ground_subtype`. As a result, we never create the constraint
to flow a `React$Element` instance to the return type of the call.

This might come up if a component class is imported from an untyped module.

Reviewed By: gabelevi

Differential Revision: D4413819

fbshipit-source-id: ad5929e91e710fd361896fc3fb8c066ac5caedef