fix returning complex types from constructors
Summary:
`_ ~> ObjTestT` was handled too early in the pattern match. for lowers
that are handled further down, like type apps and unions, we were deciding that
they were not object-like and returning the default (a new, empty object).

this diff moves the pattern further down, so that TypeAppT gets expanded, unions
get broken up, etc. I put it with similar object handling code because that
seemed to make it work and made some organizational sense. this could be one of the
very last patterns, if it was possible for everything to be last...

Reviewed By: gabelevi

Differential Revision: D4412302

fbshipit-source-id: 3c0b24813d808c991ea8c9293ab5c8525a5a57c0