Constant folding array literal rest elements
Summary:
This is a proof of concept for how I plan to revamp RestT. Rather than
solving the case that I need, this tries to solve a simplier case: spreading
tuples into array literals.

For the RestT revamp, there are a bunch of things we need to be able to do

* Assert that all rest elements are actually arrays (easy)
* Extract the ArrT type info from rest elements (harder)
* Treat tuple rest elements as multiple normal elements (hardest)

The general idea here is to defer constructing the ArrT type until we've been
able to resolve all the rest elements in an array literal. Then we can flatten
tuples into normal elements (treat `[1, ...[2, 3], 4]` like `[1, 2, 3, 4]`.

Since this IS constant folding, we need to be careful of infinite recursion.
This diff takes avikchaudhuri's advice from {D3199590}, which is to keep track of the
reasons which we've already seen and use duplicate reasons as an indicator of
being in a loop.

Reviewed By: samwgoldman

Differential Revision: D4064784

fbshipit-source-id: 3dc0bd32fa26a9b0012256e877433311c2bd5138