store errors in server env instead of globals
Summary:
eliminates the global refs that we mutate to store errors in the
server's master process. instead, saves them to the server env.

we already collate them from a map to a list and store that in the env
(`env.errorl`). i intend to get rid of `errorl` and recompute it on demand, but
this diff should be perf neutral and that one might not be, so doing them
separately.

Reviewed By: samwgoldman

Differential Revision: D4432999

fbshipit-source-id: 014f6162d10afc44a0d86cf1a64c9424b44cb861