[cleanup] pass around ErrorSets instead of error lists
Summary:
`ErrorSet` is the only thing that defines the sort order for errors.
We convert to an `error list` sooner than we have to (and in the case of parse
errors, create an `error list` directly, ignoring proper sorting).

This passes `ErrorSet`s all the way through to the printers.

Reviewed By: samwgoldman

Differential Revision: D4433464

fbshipit-source-id: 6e7d3e2b05c0bc5016f716df6caa4cc1ea6a9d96