remove reason origins (for now)
Summary:
I want to add this back soon, but for now it's not doing anything and
making error printing more complicated. Making sure that repositioning always
tracks the origin is going to be slightly harder than what this is doing (which,
AFAICT, is why this doesn't show up right now).

Reviewed By: gabelevi

Differential Revision: D4440325

fbshipit-source-id: 612741fe8f8bea091645941787017e367c865ed1