Split ObjAssignT into ObjAssign{To,From}T
Summary:
Given `Object.assign(target, source)`, `ObjAssignT` has a boolean
parameter which says whether it is currently resolving `source` or `target`.
This is kind of hacky, so I'm splitting it up. Now it goes

  target ~> ObjAssignTo
  source ~> ObjAssignFrom

I've also updated the `ObjAssign*` types to carry along whether the source is a
spread or a regular argument. We can basically treat

  Object.assign(target, ...Array<source>)

as

  Object.assign(target, ...source)

and later we can iterate through tuples.

Reviewed By: jeffmo

Differential Revision: D4405026

fbshipit-source-id: f5ce422dc81a35943781eeceec780cab411df2b3