change reasons of lookups generated by structural_subtype
Summary:
in other uses of `LookupT`, the reason refers to the object the lookup
is happening on. for instance, an `EPropAccess` error raised by an
invalid `InstanceT ~> LookupT` generates a message like `X. Covariant
property x incompatible with contravariant use in Y`, where `Y` is the reason of
the `LookupT`.

in `structural_subtype`, the `LookupT`'s reason referred to the prop instead of
the object the prop is on. it also happens to be redundant with `propref`'s
reason.

Reviewed By: gabelevi

Differential Revision: D4455129

fbshipit-source-id: 66df3588586437c4e50ca49b4b732b12c0e6b27c