Generalize EIncompatibleObject, improve invalid subtype errors
Summary:
improves the errors when A is not a subclass of B, but was supposed to
be. generalizes `EIncompatibleObject`'s name since it actually works on any two
def types that are incompatible, and uses that to thread the use_op attached to
the `ExtendsT` check through to the error message.

Reviewed By: gabelevi

Differential Revision: D4445843

fbshipit-source-id: 79b3a86279221630002c7b39e8c77c860e308494