only write_prop on literals, not all exact
Summary:
There is some logic in the `(ObjT _, ObjT _)` rule that adds optional
properties from the upper bound to the lower bound, if the lower bound is exact.

This logic predates exact object annotations, which should absolutely not be
widened (they are exact after all). Instead, this logic should be conditional on
whether the lower bound is a literal.

The logic after this change is still buggy, but less buggy as we can't violate
the core invariant of exact objects any more.

Fixes https://github.com/facebook/flow/issues/2550

Reviewed By: gabelevi

Differential Revision: D4440430

fbshipit-source-id: cf294c1bc9777a30a8de361e6a34e66625b6802c