kill weak mode
Summary:
flow weak was designed to mimic TypeScript, in that missing annotations would
be treated as any instead of implicitly static types that can be inferred.

It's not been maintained for a while, so any continued use probably detracts
from the Flow experience while not giving a good TypeScript experience either.

TODO: update docs

Reviewed By: jeffmo

Differential Revision: D4475119

fbshipit-source-id: 7d912f14155a0396fc61fa810650d34506236aa9