Carry on even if there is no TERM environment variable set.
There are a few continuous integration systems which act as STDOUT is a
TTY, but don't set `TERM`. On those occasions, we get the following
error message:

    Fatal error: exception Not_found

This, obviously, makes it pretty hard to figure out what the actual
problem is. An alternative implementation could throw a better error,
but it felt like defaulting to `"dumb"` was a reasonable solution here.