[PR] Carry on even if there is no TERM environment variable set.
Summary:
There are a few continuous integration systems which act as STDOUT is a
TTY, but don't set `TERM`. On those occasions, we get the following
error message:

    Fatal error: exception Not_found

This, obviously, makes it pretty hard to figure out what the actual
problem is. An alternative implementation could throw a better error,
but it felt like defaulting to `"dumb"` was a reasonable solution here.

I couldn't see a reasonable place to put tests, so I omitted them. If there
is, please let me know.
Closes https://github.com/facebook/flow/pull/3305

Differential Revision: D4486056

Pulled By: mroch

fbshipit-source-id: e3a7cf88d6edc49e603b18a7e7f1ba11089eb0b7