Update handling of JSXText to trim and filter empty text
Summary:
The recent support for the `// jsx Foo` pragma exposed that Flow wasn't properly desugaring JSX.  `JSXText` children are preprocessed and filtered before being passed to `React.createElement` or the custom function. Using babel as reference, this is what should happen

* Each line of the `JSXText` is trimmed of whitespace
* Empty lines are removed
* Tabs are replaces with spaces
* The lines are joined together with spaces.

Furthermore, empty `JSXText` children are thrown away and are not passed to the function. So

```
<Foo> {true} </Foo>
```

is correctly parsed as having 3 children (`" "`, `true`, and `" "`), but should desugar to `React.createElement(Foo, null, true)`;
Closes https://github.com/facebook/flow/pull/3046

Reviewed By: mroch

Differential Revision: D4346835

Pulled By: gabelevi

fbshipit-source-id: 942269157486ddebc760019e6be80e10df797229