RestT is dead, all hail the new multiflow hotness
Summary:
First off, sorry this diff is so large. It was difficult to remove `RestT` bit by bit at this point while making sure that `RestT` didn't show up in the wrong place.

== Multiflow ==

This big change in this diff is the multiflow code, which generally applies function arguments to function parameters. The way it works now is as follows:

1. Wait for spread arguments to resolve.
2. Turn tuple/array-literal spread arguments into regular arguments.
3. Any remaining spread arguments get combined into a single spread argument
4. Flow all normal arguments to normal parameters.
5. If there is a spread argument, flow it to the remaining parameters
6. Combine the remaining arguments (and spread argument if it exists) into an array literal and flow it to the rest parameter.

Step 3 is a little tricky to understand. Consider the following code

  foo(123, ...arrayOfNumber);
  bar(123, ...arrayOfNumber, ...arrayOfString);
  baz(123, ...arrayOfNumber, true, ...arrayOfString);

For the `foo()` call, we want to flow `123` to the first param and `number` to the remaining param. If we wanted to be completely sound too, we should flow `undefined` to each remaining params, since `arrayOfNumbers` might be empty. But this causes loads of errors and it's more consistent not to.

For the `bar()` call, we flow `123` to the first param and `number | string` to the remaining params. This is because we don't know the length of either array, so each parameter after the first might get a `number` or might get a `string`. So internally we'll treat it like

  bar(123, ...arrayOfNumberOrString)

for the `baz` call, we flow `123` to the first param and after that it's tricky. Param 2 may get `number` or `true`. Params 3 on may get `number`, `true`, or `string`. At the moment, the code oversimplifies the call to

  bar(123, ...arrayOfNumberOrTrueOrString)

this is an oversimplification, since param 2 will never actually be a `string`.

== Apply ==

Without any special handling, `func.apply(that, args)` would be treated like `bound_func(...args)`. However, sometimes we find code like

  func.bind.apply(null, args);

and since we don't support `func.bind(...args)` at the moment, we need special handling. Luckily, it's easy to wait for the spread to resolve and create the `CallT` once it does.

Reviewed By: samwgoldman

Differential Revision: D4442117

fbshipit-source-id: 78c2ad664b9c5ddb3db165d2b6cd9e88f65a30cf