[hack] Create Ide_api_types module
Summary:
Internally, we use Pos.t to represent all kinds of positions: single
characters, ranges, ranges in files, single characters in files.

We do this by just setting the filename to "" and ignoring it, or setting start
and end position to the same thing, etc. It doesn't really make a nice looking
API, so in previous diffs separate structures to represent each of them were
introduced.

Using them in more and more places reveals the weaknesses of "dump everything
into utils" pattern, so most of this diff is untangling it to remove
unnecessary dependencies.

Reviewed By: arxanas

Differential Revision: D4465057

fbshipit-source-id: a493790a2b042f70931942446c32482ade2da94d