non termination of calls to polymorphic functions
Summary:
Various outstanding reports of non-termination had the same root cause: we were
not caching instantiations when executing calls to polymorphic functions, which
led to cycles where different instantiations would continue to spew
different-looking, but ultimately redundant, constraints.

We already solve a similar problem with method calls on type applications. Here
we start from the same solution, but need to extend and tweak it a bit to make
it work.

Reviewed By: samwgoldman

Differential Revision: D4521320

fbshipit-source-id: e074af41546b3c4d1ad38fb181062c3f347099f6