[hack] JSON RPC version of format request/response
Summary:
The new / old versions of format are so different (different input / output), that I decided to pipe them through a separate IDE_FORMAT call. We already do something like that for AUTOCOMPLETE/IDE_AUTOCOMPLETE or FIND_REFS/IDE_FIND_REFS.

The format command is also the only one that in current implementation permits an error result - other ones either silently return an empty value, or just uncontrollably crash the entire connection.

There is a task to build out better error handling of requests that we can recover from (#15542876), and we'll likely want to build it BEFORE deploying the new API, but it's not there yet - hence I'm handling the FORMAT error case in ad hoc manner.

Reviewed By: arxanas

Differential Revision: D4523587

fbshipit-source-id: 9b15805eb83e0a6ab2f15839fcaf4f89e3de81c9