simple union opt
Summary:
It turns out that unions + subclass checking can quickly blow up, even with
small unions of classes.

Recognizing equality early is important for performance. This diff implements a
simple optimization to recognize when a union is a subset (hence a subtype) of
another union. The equality operation has to be reasonless to be effective.

Fixes https://github.com/facebook/flow/issues/3352

Reviewed By: samwgoldman

Differential Revision: D4527174

fbshipit-source-id: b942b7c35b9cceed74645608fbbcc0251df34c9b