more any propagation
Summary:
Use types that have a `tout` should any-propagate. There are a large number of
such use types which don't. This diff adds a few more.

The `UseT (_, ClassT _)` case is slightly interesting. `mk_instance` uses this
use type in order to extract an instance, so the argument to `ClassT` is a
`tout`-like thing.

Reviewed By: mroch

Differential Revision: D4548319

fbshipit-source-id: 1ae12974f0e26b8e29b95a7bd21334ea63f37c3a