kill get-importers
Summary:
get-importers is an obscure command that adds horrendous complexity to the
current implementation. Killing it also kills a lot of code, which improves the
chances of being able to clean up the rest of Module_js (most of which is
critical code and a big source of incremental bugs).

Some comments might have regressed with this diff but they're already so full of
alternative facts that they need a rewrite anyway. I'll do that in a follow-up diff.

Reviewed By: nmote

Differential Revision: D4575490

fbshipit-source-id: 28850ee1dabc08affb816c0e619b42ceb4b6bb3f