make Flow_js.reposition only reposition, not rename, reasons
Summary:
`Flow_js.reposition` took a reason and a type, and repositioned the
type to the **location** of the reason.

however, in some cases it made new tvars and used `reason` as the reason. if
anything uses `reason_of_t` on one of these tvars, it gives back `reason`
instead of the original reason, even though the reason description can be
different. `reposition` shouldn't ever change the description.

so, instead, pass a `Loc.t` instead of a `Reason.t`.

Reviewed By: avikchaudhuri

Differential Revision: D4467072

fbshipit-source-id: 636ef263b3bfb22988dfaf40871e9f0e946820dc