ReactKit: instanceOf
Summary:
React.PropTypes.instanceOf is broken, and has been since type arg arity
strictness was enforced. Before, when `for_type` was `true`, we returned an
`AnnotT`, but when we switch it to `false`, we returned a tvar with a lower
bound, which is ineffective as an upper bound.

This diff restores the `for_type` behavior and works around type arity
strictness by using ReactKit instead of calling mk_instance directly.

Reviewed By: gabelevi

Differential Revision: D4418227

fbshipit-source-id: 720e39aeb9f2e8a47917893f9ccf2bdc4b73bf8b