React propTypes should be sealed/inexact, like a type annotation
Summary:
Same as D4418280, but for the props object type we create for React.createClass
components. The generation of this type should be the same as that for shape:
both take an object mapping property names to prop validator functions, for
which we infer a type.

Like shapes, the props object is not exact w.r.t. the inferred type. It may have
arbitrarily more properties. It's also unexpected to get the shadow read/write
behavior for these objects.

Reviewed By: gabelevi

Differential Revision: D4418445

fbshipit-source-id: 6c2bb259bd729ccc2397453d55de7f710dac2c8e