ReactKit: prop types
Summary:
Before this diff, we processed the propTypes property of React.createClass
specifications through a very limited AST traversal. If the prop types were not
exactly the AST we wanted, we inferred `any`.

This diff moves that entire process out of the AST traversal phase and into the
constraint graph via a set of `CustomFunT`'s and a couple new ReactKit tools.

The first tool simplifies "complex" prop types, like `arrayOf`, by waiting on
the argument and if the argument is also a prop type, emitting the appropriate
prop type back out.

The second tool waits on the entire prop types object, then each property in
turn, and if the values are prop type custom funs, uses the inferred types
before falling back to `any`.

Reviewed By: gabelevi

Differential Revision: D4418734

fbshipit-source-id: af9ee85d503bfb3c0f95129e7f4c6e73d491278b