Remove inconsistent RReactComponentProps reason description
Summary:
When we create an instance from React.createClass calls, the default props type
is AnyObjT with a RReactComponentProps reason. Later, if we do see propTypes, we
create an ObjT with a RReactPropTypes reason.

This diff removes RReactComponentProps entirely and uses RReactPropTypes in both
places. Since the former description was used on a permissive type (Object), it
was rarely seen, so I picked the other one.

Reviewed By: mroch

Differential Revision: D4425967

fbshipit-source-id: 3fc464abdd53af768bc415d977d518eae45552e6