Position React.createClass props type at propTypes value
Summary:
Previously we would point to the location of the `propTypes` key, but that is
only possible because we have access to the full AST. In every other scenario,
the object type points to the actual type declaration. React prop types are like
an object type declaration, so I think it makes sense to point in the same
place.

Reviewed By: mroch

Differential Revision: D4426029

fbshipit-source-id: 5141680b82590bb2b7589c4f515fa049e663a6a6