Add `Bind` type destructor
Summary:
Methods defined in the spec object argument to React.createClass are autobound
to the component's `this` object. Unfortunately the pass-though flag to `BindT`
is not sufficient, because the bound types would lose their upper-bound
behavior. That is, we expect the following code to fail:

```
React.createClass({
  p: (null: ?string),
  m() {
    this.p = 0;
  },
});
```

Without a type destructor, we would create a tvar and flow `?string` into it
via `BindT`, and would end up with a tvar having `null`, `void`, and `string`
lower bounds, which would permit a new `number` inflow.

With a destructor, the `number` flows into an `EvalT` and eventually into the
`?string` upper bound, resulting in the correct error.

Reviewed By: gabelevi

Differential Revision: D4485378

fbshipit-source-id: 4a4916ca3742a1a97e577f28f78a35e7dd397a14