ReactKit: allow unsealed statics where possible
Summary:
Even though React spec objects have a `statics` key, which can be used to define
component class's static properties, it's still quite common to assign these
static properties directly to the class after the call to createClass.

This diff figures out when it's safe to leave the class's statics unsealed (safe
relative to any unsealed object, of course). In order to leave the statics
unsealed, the known statics must be exact and sealed to ensure that any added
properties will not conflict with statically unknown properties from a
possibly-wider runtime value.

Reviewed By: avikchaudhuri

Differential Revision: D4535289

fbshipit-source-id: f95550170501b4ffcce6b018e98dd0014dc10734