ReactKit: justify ignoring flags when resolving shape and propTypes (for now)
Summary:
A shape or propType spec object should be exact and sealed. If it isn't, it may
be wider at runtime than it seems and we could miss runtime errors.

However, object spread expressions create objects which are always unsealed.
It's fairly common to spread one component's propTypes spec into another's, so
we can't enforce this yet. Future work on precise object spread will unlock
this.

Reviewed By: avikchaudhuri

Differential Revision: D4535753

fbshipit-source-id: 2932c7992490995d3de2d866b4300e57794aae79