ReactKit: spec object must be exact and sealed
Summary:
We assume if a property, say `statics`, is not in the prop map of the spec
object, then it doesn't exist. However, if the spec object is inexact or
unsealed, the runtime value may have properties we don't know about, making this
analysis inaccurate.

It should be extraordinarily rare that the spec object is not a literal object,
which is exact and sealed.

Reviewed By: avikchaudhuri

Differential Revision: D4535817

fbshipit-source-id: 4807c00a9502639becab4701a7d828d0a0266e50