Log CollateErrors event
Summary:
I recently changed how we process suppressions to also consider the
locations in the extra section. I was curious if looking at more locations
would be expensive. Then I realized that we aren't profiling this part of the
code. So I'm changing that.

The good news is that collate_errors seems to be really cheap.

Reviewed By: jeffmo

Differential Revision: D4605428

fbshipit-source-id: c92bc0df0cc3bd1fc7a6b9c263abb62e8a62b9f7