error clearing on cycle break
Summary:
When a Haste cycle is broken by (say) clearing the contents of the leader of the
cycle, the resulting "Required module not found" error gets attached to a
different file in the cycle (which acts as the new leader). Unfortunately, when
the contents of the original leader was restored, we neglected to clear that
error off the intermediate leader.

This bug was observed when adding a large number of suppressions using tool,
which uses node's fs writeFile function; this caused an initial wave of updates
that cleared the contents of the involved files before writing new contents in
another wave of updates.

Reviewed By: mroch

Differential Revision: D4634848

fbshipit-source-id: dfe73ffdb6e40274af25a302fd6d30781da5aaef