eliminate dummy resolved requires for non-@flow files
Summary:
We had dummy entries in ResolvedRequiresHeap for non-@flow files. Turns out they
were only needed because (1) non-@flow files were considered "modified" for the
purposes of computing direct dependents (whose requires need to be re-resolved
when non-flow files are, e.g., added), and (2) modified files were then also
merged, causing resolved requires to be looked up for non-@flow files.

This diff keeps (1) but fixes (2) so that non-@flow files are not merged (which
was an invariant before (1) was fixed). This eliminates the need for having the
dummy entries, saving space and quite possibly also time.

Reviewed By: samwgoldman

Differential Revision: D4613584

fbshipit-source-id: c22af075ddbb7abd4f0610827147c2ed38d47d47