Object types with dictionaries should still be sealed
Summary:
Object types with an indexer properties are "unsealed" in the sense that we can
write `o.p = x` when `p` is not declared as a named property on the object, but
that interpretation doesn't really align with what `unsealed` means today: which
is that unknown reads/writes introduce "shadow" properties.

Indeed, an indexer with a number-typed key shouldn't allow a `o.p = x` write,
because "p" is definitely non-numeric.

Furthermore, because `not sealed` implies `exact`, any object with a dictionary
was considered exact as well, which isn't quite right either.

None of this mattered much until now, when we are relying on these flags as a
proxy for understanding the layout of objects (own vs. proto) in order to
calculate spread, etc.

Reviewed By: avikchaudhuri

Differential Revision: D4637975

fbshipit-source-id: e84b00097e0c34c8dc3102beef4d4cc5f010ffe8