[PR] Work around error with Result when compiling JS build
Summary:
`hack/third-party/core/result.ml` and the opam `result` module, which is a dependency of `lwt` and `base64`, which are dependencies of `js_of_ocaml`, both provide `result.cma`, leading to a Warning 31 about two modules with the same name. Warning 31 is an error in ocaml-4.03. Since this was a problem in 4.02 and things were ok, we'll just suppress the error until we can find a real solution.
Closes https://github.com/facebook/flow/pull/3444

Differential Revision: D4642365

Pulled By: mroch

fbshipit-source-id: 6b02f97fcd369bc12be97d7e19d775294a7ef821