clean up handling of file-module assocs
Summary:
This should be the final diff in this sprint (cleaning of Module_js).

The code is in a pretty happy shape now, without any funny special cases.

(1) Some files change.
(2) We look up their exported modules, and look up their providers in turn.
(3) We update the providers Hashtbl.
(4) We repick modules and update NameHeap.

Reviewed By: mroch

Differential Revision: D4671740

fbshipit-source-id: 44b0897481bf0ae16627f6b5a4b2c1a2338391c2