change Context.module_name to a local module reference
Summary:
Earlier, Context.module_name for a file pointed to the module name mapped to
that file in InfoHeap. This was redundant, but not obviously so. All that was
needed was a local reference to the exported module, like existing local
references to imported modules, to store / load corresponding tvars in
Context.module_map.

This diff localizes the scope of Context.module_name for non-lib files. For lib
files, the module name (Files.lib_module) played a slightly non-local role, so
we don't mess with it.

The main benefit of this work is that Module_js.exported_module is now called
only once per file, before CommitModules (and not during Infer).

Reviewed By: mroch

Differential Revision: D4673332

fbshipit-source-id: de8f54cf982f75d1b7b525a79910c7163f7e7121