Fix spread argument nontermination
Summary:
There were reports of nontermination on GitHub. I looked at one of the
repros, which I could distill to the smaller repro:

  function foo(...args) { foo(...args); }
  foo();

Without the `foo()` call, it was fine. However, the `foo()` call could send it
into infinite recursion. Not sure why the caching doesn't catch this case.
Anyway, there's a better malicious case of

  function foo(...args) { foo(1, ...args); }
  foo();

Where we infer `args` to be `[]`, then `[] | [1]`, then `[] | [1] | [1, 1]`,
and so on. We definitely need to bust the constant folding.

The cleanest way is to turn tuples and array literals into non-literal arrays,
which defeats the constant folding. So in the above case, `args` will
eventually be inferred to be `[] | [1] | Array<1>`

Fixes #3443

Reviewed By: samwgoldman

Differential Revision: D4728125

fbshipit-source-id: 475f29768b2bbbf87676badc87c75b321e7b671f