match failure on enums
Summary:
When comparing an enum-typed variable with some literal, we now insist that the
literal belongs to the enum. This pattern is mostly used with enums for ML-like
case analysis (either directly or with switch).

A curious effect of this approach is that it bans not only contradictory ===
checks (which lead to dead branches) but also tautological !== checks (which
lead to redundant conditionals).

This is still a baby step. Disjoint unions don't yet work (will try in a
follow-up diff). Also, we still don't recognize when the case analysis is
correctly fully exhaustive.

Reviewed By: samwgoldman

Differential Revision: D4719021

fbshipit-source-id: 72a8839e01bc9e7a11e924ca39b47399635dda7e