[hack] Implement naming for darray and varray
Summary:
Carries `darray` and `varray` through naming. This involves naming `darray` and `varray` when used as an expression, and naming `darray` and `varray` when used as a type annotation.

This diff intentionally does not introduce new typing semantics. `darray` and `varray` are translated into `array` for typing purposes. A subsequent diff will update `typing_defs.ml` and `typing.ml` to introduce typing versions of these constructs.

Reviewed By: eshrews

Differential Revision: D4689889

fbshipit-source-id: 135dd44b7152c2ab9d7e97bf51f8beb8ec443340