fix JS in prod
Summary:
when `JEKYLL_ENV=production`, it adds digests to every JS and CSS file.
this didn't integrate with requirejs.

had to move flow.js from _assets/gen to static/ to prevent it from being
digested, since the version switcher currently requires that. this is how it
worked on the old site which didn't use digests. it would be nice to fix this,
but not super urgent since the non-master files are already versioned with their
version tags.

(Note: this ignores all push blocking failures!)

Reviewed By: samwgoldman

Differential Revision: D4750803

fbshipit-source-id: d78a54e4ebc14a0c68a13697344d2caaa3d89c28