flow focus-check
Summary:
Experimental command that runs a focused check on a target file.

Currently, the run still does a lot of work.

(1) All files are parsed.
(2) Local inference is run on all files.
(3) Recursive dependencies of the target file are computed.
(4) Global merge is run only on those dependencies.

Only (4) is something that we *have* to do.

Immediate follow-up work:

* In (2), run local inference only on the dependencies computed by (3).
* Optimize (3), since recursive dependencies are computed anyway during (4).

Further out:

* Only parse the dependencies computed by (3). Note that the headers of all
  files must be scanned in any case for module dependency management (although
  the design of the module system could help relax this assumption).
* Also check the dependents of the target file.
* Eventually implement force-status, which keeps the server running and
  incrementally widens the set of files that are type checked.

Reviewed By: samwgoldman

Differential Revision: D4680021

fbshipit-source-id: 44117c6cd9016c3966ba18ec254351ef1b87392c