part 2: streaming infer
Summary:
Previously, we would infer all files before merging only those files that are
dependencies of the file we're focusing on.

Now, infer will happen in a streaming manner, much like merge already
happens. In particular, we will add dependencies to the queue as they are
discovered, so effectively we will only infer those files that are dependencies
of the file we're focusing on.

A nice effect of this change (besides big improvements in perf) is that now we
won't report errors in unrelated files.

Reviewed By: samwgoldman

Differential Revision: D4760270

fbshipit-source-id: 8893d06279a766bd5997d2295f362ae06da2971b