Defer spread until type parameters are substituted
Summary:
Before this change, we evaluated the result of a type spread eagerly, when we
visited the annotation. This diff changes that to instead treat the spread as a
term itself, which is evaluated when "used."

Because type spreads are type terms, they can contain bounds which are
substituted the normal way. This means we no longer output an internal "Did not
expect BoundT" error when generics appear inside type spreads.

However, like other type terms, the result might never be evaluated, such that
some errors-by-construction like exact-spread-of-inexact may not be realized.
This is an example of a larger issue where type terms should be evaluated once
all bound types are substituted out and no later.

Note that this does not fully solve the problem of generic type spread! Flow
will instantiate type parameters to `empty` and `mixed`, which are both
unsupported by the current implementation. Adding full support for generic
spread is still a TODO, but poses significant additional complexity.

Ref https://github.com/facebook/flow/issues/3535

Reviewed By: gabelevi

Differential Revision: D4757510

fbshipit-source-id: 22039a33ac78d2d41e907ace45a0cfa1a0e39e98