[PR] Add documentation for $ObjMap
Summary:
This is long due.

I volunteered [some time ago](https://github.com/facebook/flow/issues/2464#issuecomment-274808633) to document advanced features, and thejameskyle was so nice to even reach out and ask for help for the new website that saw the light yesterday.

Unfortunately "life happened" and I couldn't find the time to contribute as I thought.

Now that the new gorgeous docs are out, I feel really sorry for not having helped like I promised and I think the minimum I can do is to fill some of the remaining the gaps.

Here's the documentation for `$ObjMap`, a powerful advanced feature that many asked to explain.

Hope this helps!
Closes https://github.com/facebook/flow/pull/3589

Differential Revision: D4788146

Pulled By: samwgoldman

fbshipit-source-id: 3060ac99d404a1d1d4f47ec0768c97642c3037e5