[PR] Fix pre-publish command in install guide (fixes #3594)
Summary:
I've broken the pre-publish command install guide in #3594 (now the package manager is completely missing).

This fixes it (this time I tested the website locally to ensure it works!)

Bonus: I've also added the same fix to the flow-remove-type section, so that now it respects the `yarn`/`npm` selection too.

![](http://g.recordit.co/A5zugGrBgu.gif)
Closes https://github.com/facebook/flow/pull/3598

Differential Revision: D4796681

Pulled By: samwgoldman

fbshipit-source-id: 2de4faab7040456dd55f15d07141a16d7fb4c0d9